Add more information to exceptions
Motivation:

When an exception is thrown we don't have much information about it.

Modifications:

When getting an exception, first rethrow to get the stack trace on the
pg lib. Then add more info, such as the query that originated it and
fields that might be present.

Result:

We can now get valuable information in Bugsnag.