Fix rollback issues
Motivation:

We had several issues with the rollback:

* Clients of `transaction-run!` were expecting a result channel, but if
  we had an error, we'd return a result
* If the rollback failed (and it does by what apperars to be a bug on
  the driver), no result or channel was returned, but an exception was
  raised

Modifications:

The `rollback` fn now has an inner try-catch. If the call succeeds, a
new channel will be returned. If the call fails (and that was our
problem), the exception is catched, and wrapped by a channel with the
result. It will also be logged.

So this means that we can have a rollback that failes for some reason,
and we return a result. Those errors will be catalogued and should be
verified one by one.

On this scenario we had a problem where the socket was closed and we
created an issue with the driver.